ml/ml_requirements.txt
torch==1.9.0
numpy==1.20.3

# dev
matplotlib==3.4.2 
scipy==1.7.0
pytest==6.2.4
pytest-django==4.4.0
